Added logic for login/register and shopping cart window

+++ Project-Work/src/Store/authSlice.js
@@ -6,6 +6,7 @@ const authSlice = createSlice({
         isSignUp: false,
         showPassword: false,
         users: JSON.parse(localStorage.getItem('users')) || [],
+        currentUser: null,
     },
     reducers: {
         toggleSignUp: (state) => {
@@ -15,18 +16,41 @@ const authSlice = createSlice({
             state.showPassword = !state.showPassword;
         },
         signIn: (state, action) => {
+            console.log('Users from localStorage:', JSON.parse(localStorage.getItem('users')));
 
+            const userData = action.payload;
+
+            const user = state.users.find(user => user.Email === userData.Email && user.Password === userData.Password);
+
+            if (user) {
+                state.currentUser = user;
+                console.log('Hello');
+            } else {
+                console.error('Invalid email or password');
+            }
         },
         signUp: (state, action) => {
             const newUser = action.payload;
 
-            state.users.push(newUser);
-            localStorage.setItem('users', JSON.stringify(state.users));
+            if (!state.users.find(user => user.Email === newUser.Email) && newUser.Password === newUser.ConfirmPassword) {
+                state.users.push({
+                    UserName: newUser.UserName,
+                    Email: newUser.Email,
+                    Password: newUser.Password
+                });
+
+                localStorage.setItem('users', JSON.stringify(state.users));
+            } else {
+                console.error('User already exists or passwords do not match');
+            }
+        },
+        signOut: (state) => {
+            state.currentUser = null;
         }
     },
 });
 
-export const { toggleSignUp, togglePasswordVisibility, signUp } = authSlice.actions;
+export const { toggleSignUp, togglePasswordVisibility, signIn, signUp, signOut } = authSlice.actions;
 
 export const selectAuth = (state) => state.auth;
